# backend/apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "香港诊所保险保单智能填报 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/templates": {
      "post": {
        "tags": [
          "admin:templates"
        ],
        "summary": "Upload Template",
        "operationId": "upload_template_api_admin_templates_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_template_api_admin_templates_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "admin:templates"
        ],
        "summary": "List Templates",
        "operationId": "list_templates_api_admin_templates_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateOut"
                  },
                  "title": "Response List Templates Api Admin Templates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/templates/{template_id}": {
      "get": {
        "tags": [
          "admin:templates"
        ],
        "summary": "Get Template",
        "operationId": "get_template_api_admin_templates__template_id__get",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/templates/{template_id}/fields": {
      "get": {
        "tags": [
          "admin:templates"
        ],
        "summary": "List Fields",
        "operationId": "list_fields_api_admin_templates__template_id__fields_get",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateFieldOut"
                  },
                  "title": "Response List Fields Api Admin Templates  Template Id  Fields Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin:templates"
        ],
        "summary": "Create Field",
        "operationId": "create_field_api_admin_templates__template_id__fields_post",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Template Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateFieldCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateFieldOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/templates/{template_id}/fields/{field_id}": {
      "put": {
        "tags": [
          "admin:templates"
        ],
        "summary": "Update Field",
        "operationId": "update_field_api_admin_templates__template_id__fields__field_id__put",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Template Id"
            }
          },
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateFieldUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateFieldOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin:templates"
        ],
        "summary": "Delete Field",
        "operationId": "delete_field_api_admin_templates__template_id__fields__field_id__delete",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Template Id"
            }
          },
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Field Api Admin Templates  Template Id  Fields  Field Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/templates/{template_id}/preview-fill": {
      "post": {
        "tags": [
          "admin:templates"
        ],
        "summary": "Preview Fill",
        "operationId": "preview_fill_api_admin_templates__template_id__preview_fill_post",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Template Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewFillRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewFillResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/templates/{template_id}/publish": {
      "post": {
        "tags": [
          "admin:templates"
        ],
        "summary": "Publish Template",
        "operationId": "publish_template_api_admin_templates__template_id__publish_post",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/doctor/insurance-companies/{company_id}/templates": {
      "get": {
        "tags": [
          "doctor:claims"
        ],
        "summary": "List Templates",
        "operationId": "list_templates_api_doctor_insurance_companies__company_id__templates_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateBrief"
                  },
                  "title": "Response List Templates Api Doctor Insurance Companies  Company Id  Templates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_template_api_admin_templates_post": {
        "properties": {
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "template_name": {
            "type": "string",
            "title": "Template Name"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "template_name",
          "file"
        ],
        "title": "Body_upload_template_api_admin_templates_post"
      },
...
      "FieldMappingOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "standard_field_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standard Field Id"
          },
          "transform_rule_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transform Rule Id"
          },
          "fixed_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fixed Value"
          },
          "checkbox_map_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkbox Map Value"
          }
        },
        "type": "object",
        "required": [
          "id",
          "standard_field_id",
          "transform_rule_id",
          "fixed_value",
          "checkbox_map_value"
        ],
        "title": "FieldMappingOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreviewFillRequest": {
        "properties": {
          "sample_values": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Sample Values"
          }
        },
        "type": "object",
        "required": [
          "sample_values"
        ],
        "title": "PreviewFillRequest"
      },
      "PreviewFillResponse": {
        "properties": {
          "preview_pdf_url": {
            "type": "string",
            "title": "Preview Pdf Url"
          }
        },
        "type": "object",
        "required": [
          "preview_pdf_url"
        ],
        "title": "PreviewFillResponse"
      },
...
      "TemplateBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "template_name": {
            "type": "string",
            "title": "Template Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "template_name",
          "version",
          "page_count"
        ],
        "title": "TemplateBrief"
      },
      "TemplateFieldCreate": {
        "properties": {
          "page_no": {
            "type": "integer",
            "title": "Page No",
            "default": 1
          },
          "field_label_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Label Raw"
          },
          "field_type": {
            "type": "string",
            "title": "Field Type",
            "default": "text"
          },
          "pos_x": {
            "type": "number",
            "title": "Pos X"
          },
          "pos_y": {
            "type": "number",
            "title": "Pos Y"
          },
          "width": {
            "type": "number",
            "title": "Width"
          },
          "height": {
            "type": "number",
            "title": "Height"
          },
          "font_size": {
            "type": "number",
            "title": "Font Size",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "pos_x",
          "pos_y",
          "width",
          "height"
        ],
        "title": "TemplateFieldCreate"
      },
      "TemplateFieldOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "template_id": {
            "type": "integer",
            "title": "Template Id"
          },
          "page_no": {
            "type": "integer",
            "title": "Page No"
          },
          "field_label_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Label Raw"
          },
          "pdf_field_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Field Name"
          },
          "field_type": {
            "type": "string",
            "title": "Field Type"
          },
          "pos_x": {
            "type": "number",
            "title": "Pos X"
          },
          "pos_y": {
            "type": "number",
            "title": "Pos Y"
          },
          "width": {
            "type": "number",
            "title": "Width"
          },
          "height": {
            "type": "number",
            "title": "Height"
          },
          "font_size": {
            "type": "number",
            "title": "Font Size"
          },
          "recognize_source": {
            "type": "string",
            "title": "Recognize Source"
          },
          "confidence_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "is_confirmed": {
            "type": "boolean",
            "title": "Is Confirmed"
          },
          "row_version": {
            "type": "integer",
            "title": "Row Version"
          },
          "mapping": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FieldMappingOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "template_id",
          "page_no",
          "field_label_raw",
          "pdf_field_name",
          "field_type",
          "pos_x",
          "pos_y",
          "width",
          "height",
          "font_size",
          "recognize_source",
          "confidence_score",
          "is_confirmed",
          "row_version"
        ],
        "title": "TemplateFieldOut"
      },
      "TemplateFieldUpdate": {
        "properties": {
          "row_version": {
            "type": "integer",
            "title": "Row Version"
          },
          "page_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page No"
          },
          "field_label_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Label Raw"
          },
          "field_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Type"
          },
          "pos_x": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pos X"
          },
          "pos_y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pos Y"
          },
          "width": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "font_size": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Font Size"
          },
          "is_confirmed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Confirmed"
          }
        },
        "type": "object",
        "required": [
          "row_version"
        ],
        "title": "TemplateFieldUpdate"
      },
      "TemplateOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "template_name": {
            "type": "string",
            "title": "Template Name"
          },
          "template_code": {
            "type": "string",
            "title": "Template Code"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "original_pdf_url": {
            "type": "string",
            "title": "Original Pdf Url"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "page_width": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Width"
          },
          "page_height": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Height"
          },
          "parse_status": {
            "type": "string",
            "title": "Parse Status"
          },
          "parse_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parse Error"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "template_name",
          "template_code",
          "version",
          "original_pdf_url",
          "page_count",
          "page_width",
          "page_height",
          "parse_status",
          "parse_error",
          "is_active",
          "created_at"
        ],
        "title": "TemplateOut"
      },
      "TemplateUploadResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "parse_status": {
            "type": "string",
            "title": "Parse Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "parse_status"
        ],
        "title": "TemplateUploadResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}